マイグレーションの修正（外部キーの設定）

+++ database/migrations/2021_02_06_130834_create_employees_table.php
@@ -6,28 +6,17 @@ use Illuminate\Support\Facades\Schema;
 
 class CreateEmployeesTable extends Migration
 {
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
     public function up()
     {
         Schema::create('employees', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('employee_id');
+            $table->unsignedBigInteger('employee_id');
             $table->string('employee_name');
             $table->string('office');
-            //$table->string('goods_id');
             $table->timestamps();
         });
     }
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
     public function down()
     {
         Schema::dropIfExists('employees');
